third_party/blink/web_tests/fast/canvas/canvas-createImageBitmap-from-canvas-toBlob-expected.txt

Tests that createImageBitmap from a canvas.toBlob object should have the same pixel data as the blob.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS image data from the created ImageBitmap and the originated blob is the same
PASS successfullyParsed is true

TEST COMPLETE
